Valid setup of zOSMF mock
