+++ config/local/api-defs/zosmf.yml
@@ -1,0 +1,26 @@
+services:
+    - serviceId: zosmf  # unique lowercase ID of the service
+      catalogUiTileId: zosmf  # ID of the API Catalog UI tile (visual grouping of the services)
+      title: Statically Defined API Service  # Title of the service in the API catalog
+      description: Mock ZOSMF service  # Description of the service in the API catalog
+      instanceBaseUrls:  # list of base URLs for each instance
+          - https://localhost:10013/  # scheme://hostname:port/contextPath
+      homePageRelativeUrl:  # Normally used for informational purposes for other services to use it as a landing page
+      routes:
+          - gatewayUrl: api # [api/ui/ws]/v{majorVersion}
+            serviceRelativeUrl:  # relativePath that is added to baseUrl of an instance
+          # List of APIs provided by the service (currently only one is supported):
+      authentication:
+          scheme: zosmf     # This service expects z/OSMF LTPA token
+      apiInfo:
+          - apiId: com.ibm.zosmf
+            gatewayUrl: api
+            version: 2.4.0
+            documentationUrl: https://www.ibm.com/support/knowledgecenter/en/SSLTBW_2.4.0/com.ibm.zos.v2r4.izua700/IZUHPINFO_RESTServices.htm
+
+
+# List of tiles that can be used by services defined in the YAML file:
+catalogUiTiles:
+    zosmf:
+        title: zOSMF
+        description: zOSMF
