+++ mock-zosmf/src/main/java/org/zowe/apiml/client/api/InfoController.java
@@ -1,0 +1,48 @@
+/*
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ */
+package org.zowe.apiml.client.api;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
+
+import javax.servlet.http.Cookie;
+import javax.servlet.http.HttpServletResponse;
+
+@RestController
+public class InfoController {
+
+    @RequestMapping(value = "/zosmf/info", produces = "application/json; charset=utf-8", method = RequestMethod.GET)
+    public ResponseEntity<?> info(HttpServletResponse response) {
+        Cookie ltpaToken = new Cookie("LtpaToken2", "paMypL7yRO/IBroQtro21/uSC2LTrJvOuYebHaPc6JAUNWQ7lEHHt1l3CYeXa/nP6aKLFHTuyWy3qlRXvt10PjVdVl+7Q+wavgIsro7odz+PvTaJBp/+r0AH+DHYcdZikKe8dytGYZRH2c2gw8Gv3PliDIMd1iPEazY4HeYTU5VCFM5cBJkeIoTXCfL5ud9wTzrkY2c4h1PQPtx+hYCF4kEpiVkqIypVwjQLzWdJGV1Ihz7NqH/UU9MMJRXY1xMqsWZSibs2fX5MVK77dnyBrNYjVXA7PqYL6U/v5/1UCvuYQ/iEU9+Uy95J+xFEsnTX");
+
+        ltpaToken.setSecure(true);
+        ltpaToken.setHttpOnly(true);
+        ltpaToken.setPath("/");
+
+        response.addCookie(ltpaToken);
+
+        return new ResponseEntity<>("{\n" +
+            "  \"zos_version\": \"04.27.00\",\n" +
+            "  \"zosmf_port\": \"1443\",\n" +
+            "  \"zosmf_version\": \"27\",\n" +
+            "  \"zosmf_hostname\": \"usilca32.lvn.broadcom.net\",\n" +
+            "  \"plugins\": {\n" +
+            "    \"msgId\": \"IZUG612E\",\n" +
+            "    \"msgText\": \"IZUG612E\"\n" +
+            "  },\n" +
+            "  \"zosmf_saf_realm\": \"SAFRealm\",\n" +
+            "  \"zosmf_full_version\": \"27.0\",\n" +
+            "  \"api_version\": \"1\"\n" +
+            "}", HttpStatus.OK);
+    }
+}
