+++ mock-zosmf/src/main/java/org/zowe/apiml/client/service/AuthenticationService.java
@@ -1,0 +1,19 @@
+/*
+ * This program and the accompanying materials are made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution, and is available at
+ * https://www.eclipse.org/legal/epl-v20.html
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Copyright Contributors to the Zowe Project.
+ */
+package org.zowe.apiml.client.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class AuthenticationService {
+    // Authenticate via header
+    // Authenticate via body
+    // Authenticate in other ways.
+}
