+++ package.json
@@ -6,8 +6,8 @@
         "doc": "docs"
     },
     "scripts": {
-        "api-layer": "concurrently --names \"GS,DS,AC,DC\" -c cyan,yellow,white,blue npm:gateway-service npm:discovery-service npm:api-catalog-service npm:discoverable-client",
-        "api-layer-ci": "concurrently --names \"GS,DS,AC,DC\" -c cyan,yellow,white,blue npm:gateway-service-ci npm:discovery-service npm:api-catalog-service npm:discoverable-client",
+        "api-layer": "concurrently --names \"GS,DS,AC,DC,ZO\" -c cyan,yellow,white,blue,green npm:gateway-service npm:discovery-service npm:api-catalog-service npm:discoverable-client npm:mock-zosmf",
+        "api-layer-ci": "concurrently --names \"GS,DS,AC,DC,ZO\" -c cyan,yellow,white,blue,green npm:gateway-service-ci npm:discovery-service npm:api-catalog-service npm:discoverable-client npm:mock-zosmf",
         "api-layer-core": "concurrently --names \"GW,DS,AC\" -c cyan,yellow,white npm:gateway-service npm:discovery-service npm:api-catalog-service",
         "api-layer-without-gateway": "concurrently --names \"DS,AC,DC\" -c yellow,white,blue npm:discovery-service npm:api-catalog-service npm:discoverable-client",
         "api-layer-without-discovery": "concurrently --names \"GW,AC,DC\" -c cyan,white,blue npm:gateway-service npm:api-catalog-service npm:discoverable-client",
@@ -29,6 +29,7 @@
         "discovery-service-2": "java -jar discovery-service/build/libs/discovery-service.jar --spring.profiles.active=https --spring.config.additional-location=file:./config/local-multi/discovery-service-2.yml",
         "api-catalog-service-1": "java -jar api-catalog-services/build/libs/api-catalog-services.jar --spring.config.additional-location=file:./config/local-multi/api-catalog-service.yml",
         "discoverable-client-1": "java -jar discoverable-client/build/libs/discoverable-client.jar --spring.config.additional-location=file:./config/local-multi/discoverable-client.yml",
+        "mock-zosmf": "java -jar mock-zosmf/build/libs/mock-zosmf.jar",
         "test": "./gradlew runAllIntegrationTests",
         "//": "echo \"for windows: to install sudo, run chocolatey.exe install sudo\"",
         "register-certificates-win": "echo \"for windows: to install sudo, run chocolatey.exe install sudo\" & sudo . cmd /C \"certutil -enterprise -delstore Root \"Zowe Development Instances Certificate Authority\"  &  certutil -enterprise -f -v -AddStore Root \"keystore/local_ca/localca.cer\" || pause\""
